+++ src/App.css
@@ -1,5 +1,6 @@
 .App {
   width: 100%;
+  font-family: 'Libre Baskerville', serif;
 }
 .App-header {
   background-color: #282c34;
